+++ frontend/src/services/parecerAnalistaGeo.service.js
@@ -1,0 +1,12 @@
+var ParecerAnalistaGeo = function(request, config) {
+
+	this.findParecerByIdHistoricoTramitacao = function(id) {
+
+		return request
+			.get(config.BASE_URL() + 'parecer/analistaGeo/' + id);
+
+    };
+
+};
+
+exports.services.ParecerAnalistaGeo = ParecerAnalistaGeo;